refactored active panel
