+++ src/app/containers/post/activePanel/index.js
@@ -1,0 +1,2 @@
+export { default } from './ActivePanel.connect';
+export { ActivePanel } from './ActivePanel';
